refactor(main): move remaining entry related code (#2127)
