+++ docs/source/source_code/src/globals.rst
@@ -1,0 +1,5 @@
+globals
+=======
+
+.. doxygenfile:: globals.h
+   :allow-dot-graphs:
