+++ src/confighttp.cpp
@@ -30,9 +30,9 @@
 #include "confighttp.h"
 #include "crypto.h"
 #include "file_handler.h"
+#include "globals.h"
 #include "httpcommon.h"
 #include "logging.h"
-#include "main.h"
 #include "network.h"
 #include "nvhttp.h"
 #include "platform/common.h"
