+++ src/platform/linux/audio.cpp
@@ -4,6 +4,7 @@
  */
 #include <bitset>
 #include <sstream>
+#include <thread>
 
 #include <boost/regex.hpp>
 
@@ -15,7 +16,6 @@
 
 #include "src/config.h"
 #include "src/logging.h"
-#include "src/main.h"
 #include "src/thread_safe.h"
 
 namespace platf {
