+++ src/platform/linux/input.cpp
@@ -20,11 +20,11 @@
 #include <cmath>
 #include <cstring>
 #include <filesystem>
+#include <thread>
 
 #include "src/config.h"
 #include "src/input.h"
 #include "src/logging.h"
-#include "src/main.h"
 #include "src/platform/common.h"
 #include "src/utility.h"
