+++ src/platform/linux/kmsgrab.cpp
@@ -13,9 +13,9 @@
 #include <xf86drmMode.h>
 
 #include <filesystem>
+#include <thread>
 
 #include "src/logging.h"
-#include "src/main.h"
 #include "src/platform/common.h"
 #include "src/round_robin.h"
 #include "src/utility.h"
