+++ src/platform/windows/publish.cpp
@@ -14,7 +14,6 @@
 #include "misc.h"
 #include "src/config.h"
 #include "src/logging.h"
-#include "src/main.h"
 #include "src/network.h"
 #include "src/nvhttp.h"
 #include "src/platform/common.h"
