+++ src/stream.cpp
@@ -18,9 +18,9 @@ extern "C" {
 }
 
 #include "config.h"
+#include "globals.h"
 #include "input.h"
 #include "logging.h"
-#include "main.h"
 #include "network.h"
 #include "stat_trackers.h"
 #include "stream.h"
