+++ src/video.cpp
@@ -19,9 +19,9 @@ extern "C" {
 
 #include "cbs.h"
 #include "config.h"
+#include "globals.h"
 #include "input.h"
 #include "logging.h"
-#include "main.h"
 #include "nvenc/nvenc_base.h"
 #include "platform/common.h"
 #include "sync.h"
